Deploy http call moved to admin section

+++ cli/update.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-PROJECT_DIR=$(realpath $(dirname $0)/../)
+PROJECT_DIR=$(dirname $0)/../
+test -d $PROJECT_DIR || exit 0
 cd $PROJECT_DIR
 git pull origin develop
 rm -rf ./var/compiled
@@ -8,6 +9,6 @@ rm -rf ./var/db_info
 rm -rf ./var/default
 rm -rf ./var/fb_cache
 rm -rf ./var/sessions
-mkdir ./var/logs
-./lib/limb/limb migrate_run
-./lib/limb/limb od_calc_interest
+test -d ./var/logs || mkdir ./var/logs
+./limb migrate_run
+./limb od_calc_interest
